+++ src/resonaate/tasking/metrics/target.py
@@ -1,11 +1,12 @@
-"""Defines behavior-focused tasking metrics."""
+"""Defines target-focused tasking metrics."""
 from __future__ import annotations
 
 # Standard Library Imports
 from typing import TYPE_CHECKING
 
 # Local Imports
-from .metric_base import BehaviorMetric
+from ...physics.constants import DAYS2SEC
+from .metric_base import TargetMetric
 
 if TYPE_CHECKING:
     # Local Imports
@@ -13,19 +14,19 @@ if TYPE_CHECKING:
     from ...agents.sensing_agent import SensingAgent
 
 
-class TimeSinceObservation(BehaviorMetric):
-    """Time since observation behavior metric."""
+class TimeSinceObservation(TargetMetric):
+    """Time since observation target metric."""
 
     def calculate(self, estimate_agent: EstimateAgent, sensor_agent: SensingAgent) -> float:
-        """Calculate the total minutes since the last observation of this target.
+        """Calculate the total seconds since the last observation of this target.
 
         Args:
             estimate_agent (:class:`.EstimateAgent`): estimate agent for which this metric is being calculated
             sensor_agent (:class:`.SensorAgent`): sensor agent for which this metric is being calculated
 
         Returns:
-            ``float``: Time since last observation in minutes
+            ``float``: Time since last observation in seconds
         """
         current_time = estimate_agent.julian_date_epoch
         observation_time = estimate_agent.last_observed_at
-        return float((current_time - observation_time) * 24 * 60)
+        return float((current_time - observation_time) * DAYS2SEC)